requirements.txt
# Python dependencies for Zoom Rooms SDK wrapper microservice
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6

# For binding generation
libclang>=16.0.0
jinja2>=3.1.0
